+++ src/main/java/com/geNAZt/RegionShop/Interface/Shop/ShopBuy.java
@@ -152,8 +152,8 @@ public class ShopBuy extends ShopCommand {
                         plugin.getDatabase().delete(item);
                     }
 
-                    Transaction.generateTransaction(player, ShopTransaction.TransactionType.BUY, region, owner.getName(), item.getItemID(), wishAmount, item.getSell(), 0);
-                    Transaction.generateTransaction(owner, ShopTransaction.TransactionType.SELL, region, player.getName(), item.getItemID(), wishAmount, item.getSell(), 0);
+                    Transaction.generateTransaction(player, ShopTransaction.TransactionType.BUY, region, owner.getName(), item.getItemID(), wishAmount, item.getSell().doubleValue(), 0.0);
+                    Transaction.generateTransaction(owner, ShopTransaction.TransactionType.SELL, region, player.getName(), item.getItemID(), wishAmount, item.getSell().doubleValue(), 0.0);
                 } else {
                     player.sendMessage(Chat.getPrefix() + ChatColor.RED +  "You have not enough money for this. You need "+ (((float)wishAmount / (float)item.getUnitAmount()) * (float)item.getSell()) + "$");
                     return;
